Fix busy loop in Iterator::forever with mio-support feature
The problem was, when the feature was enabled, the wakeup socket got
switched to non-blocking mode (no matter if anything mio-related was
actually used). So things that should have been blocking didn't and it
was retrying the read from the pipe in a tight loop.

Now we don't switch to non-blocking, but use the MSG_DONTWAIT flag when
we need non-blocking read.

Closes #16