Refactor iterator into a generic backend and blocking implementation
This is a preparation for implementing async support in their own crates.
The backend uses generic types for the reading and writting ends of
the pipe. This allows using the IO types from asynchronous runtimes
in the different async crate implemenations.

The blocking implementation contains the Signal struct using a blocking
UnixStream and provides the same API like the previous version of this
struct. So clients of this struct do not have to change.

This commit also removes the tokio and mio features from the iterator
module. The asynchronous support will be added back in their own
crates in a later commit.